Moved models from data to core module model
