+++ core/data/build.gradle.kts
@@ -31,6 +31,7 @@ android {
 
 dependencies {
     implementation(project(":core:network"))
+    implementation(project(":core:model"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.kotlinx.coroutines.android)
