+++ core/model/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    alias(libs.plugins.org.jetbrains.kotlin.jvm)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+}
