+++ core/model/src/main/java/me/brisson/revam/core/model/WallpaperListItem.kt
@@ -1,4 +1,4 @@
-package me.brisson.revamp.core.data.model
+package me.brisson.revam.core.model
 
 data class WallpaperListItem(
     val id: String,
