+++ settings.gradle.kts
@@ -17,3 +17,4 @@ rootProject.name = "Revamp"
 include(":app")
 include(":core:network")
 include(":core:data")
+include(":core:model")
